Incredibly janky way to get script to work for both unmodeled and modeled Oxcal .json files.
I would prefer to be able to def the Medians, Probabilities, Ranges, and Row_Shift codes sections once, but with my current understanding it would require passing in far to many variables to make sense.

Perhaps a code rewrite that doesn't depend on so much of the same varibles or get the code sections to be more self-contained and not need to interact with as many external variables.